requirement.txt
fastapi==0.82.0
uvicorn==0.18.3
starlette==0.19.1
pydantic==1.10.2
loguru==0.6.0
python-dotenv==0.21.0
google-cloud-bigquery==3.3.2
pandas==1.3.5
db-dtypes==1.0.3
# gunicorn==20.1.0